// src/models/AssessmentCategory.ts
export enum AssessmentCategory {
  FINANCIAL = 'FINANCIAL',
  OPERATIONS = 'OPERATIONS',
  PATIENT_CARE = 'PATIENT_CARE',
  TECHNOLOGY = 'TECHNOLOGY',
  COMPLIANCE = 'COMPLIANCE',
  FACILITIES = 'FACILITIES',
  MARKETING = 'MARKETING',
  GEOGRAPHY = 'GEOGRAPHY',
  STAFFING = 'STAFFING',
  AUTOMATION = 'AUTOMATION'
}